+++ client/src/components/Application.js
@@ -44,11 +44,7 @@ function Application() {
             {user ?
                 <Router>
                     <Switch>
-                        <Route exact path="/">
-                            {/* <Navbar /> */}
-                            <Overworld />
-                            {/* <NoMatch /> */}
-                        </Route>
+
           {/* the next 3 routes are for development purposes only. take them out later! */}
                         <Route exact path="/overworld">  
                             <Overworld />
