+++ client/src/components/actor/index.jsx
@@ -1,9 +1,9 @@
-import React, { useState, useContext } from "react";
+import React, { /*useState,*/ useContext } from "react";
 import CharContext from "../../utils/CharContext"
 import Sprite from "../sprite";
 
 
-export default function Actor({sprite, data, step = 0, dir =0}) {
+export default function Actor({/*sprite,*/ data, step = 0, dir =0}) {
     const { h, w } = data;
     const {characterState} = useContext(CharContext)
     return (
